requirements.txt

nltk==3.8.1
pyparsing==3.1.1
PyYAML==6.0.1
requests==2.31.0
wordcloud==1.9.2
